requirements.txt
openai
python-dotenv
python-dateutil
rapidfuzz
jupyter
